Move strain wrangling function to its own file in better folder
